[update] 連絡先詳細画面で使用するAPI を連絡先一覧画面で使用しており、エラーが出ていたため削除
[タスク] チュートリアルをやってみる #1
