+++ app/routes/contacts.$contactId_.edit.tsx
@@ -2,15 +2,6 @@ import { ActionFunctionArgs } from "@remix-run/node";
 import { redirect } from "@remix-run/react";
 import invariant from "tiny-invariant";
 import { updateContact } from "~/data";
-
-export const action = async ({ params, request }: ActionFunctionArgs) => {
-    invariant(params.contactId, "Missing contactId param");
-    const formData = await request.formData();
-    const updates = Object.fromEntries(formData);
-    await updateContact(params.contactId, updates);
-    return redirect(`/contacts/${params.contactId}`);
-};
-
 import type { LoaderFunctionArgs } from "@remix-run/node";
 import { json } from "@remix-run/node";
 import { Form, useLoaderData } from "@remix-run/react";
@@ -26,9 +17,19 @@ export const loader = async ({
   if (!contact) {
     throw new Response("Not Found", { status: 404 });
   }
-  
+
   return json({ contact });
 };
+
+export const action = async ({ params, request }: ActionFunctionArgs) => {
+    invariant(params.contactId, "Missing contactId param");
+    const formData = await request.formData();
+    const updates = Object.fromEntries(formData);
+
+    await updateContact(params.contactId, updates);
+    return redirect(`/contacts/${params.contactId}`);
+};
+
 
 const EditContact = () => {
   const { contact } = useLoaderData<typeof loader>();
